Rearrange media mod, create separate file for mic and decoding
